Check if requirements file exists

+++ cage/container/handler.py
@@ -88,7 +88,10 @@ class ContainerHandler:
         self.__write_to_dockerfile("COPY {} /usr/src/app".format(path))
 
     def install_deps(self, requirements_file):
-        self.__write_to_dockerfile("RUN pip install --no-cache-dir -r {}".format(requirements_file))
+        if os.path.exists(requirements_file):
+            self.__write_to_dockerfile("RUN pip install --no-cache-dir -r {}".format(requirements_file))
+        else:
+            raise FileNotFoundError(requirements_file)
 
     def redirect_logs(self, container):
         logs = self.__client.logs(container, stream=True)
